requirements.txt
Flask_Login==0.5.0
Flask_RESTful==0.3.7
Flask_Bcrypt==0.7.1
SQLAlchemy==1.3.16
Flask_SQLAlchemy==2.4.1
google_api_python_client==1.10.0
PyYAML==5.3.1
